[jit] define equality for IValue
Previously we were reluctant to define equality for IValues, as it's not
totally straightforward. But the vacuum this created basically forced
people to define their own equality comparisons for their own purposes.
We have at least 3 in PyTorch itself, and 2 others outside that I know
of.

These implementations are generally wrong, so we should just bite the
bullet and define equality canonically.

ghstack-source-id: a5cce3738756bfb7baf25c83a80020fc9abe9675
Pull Request resolved: https://github.com/pytorch/pytorch/pull/34986